Class_Assignment/multi_Agent/requirements.txt
chainlit
uv
python-dotenv
google-generativeai
openai
